Added lore + fixes (not tested)

+++ src/main/java/by/dero/gvh/model/itemsinfo/DamagePotionInfo.java
@@ -3,14 +3,14 @@ package by.dero.gvh.model.itemsinfo;
 import by.dero.gvh.model.ItemInfo;
 
 public class DamagePotionInfo extends ItemInfo {
-    private int damage;
+    private double damage;
     private double radius;
 
-    public int getDamage() {
+    public double getDamage() {
         return damage;
     }
 
-    public void setDamage(int damage) {
+    public void setDamage(double damage) {
         this.damage = damage;
     }
