+++ src/main/java/by/dero/gvh/model/itemsinfo/DragonFlyInfo.java
@@ -5,7 +5,7 @@ import by.dero.gvh.model.ItemInfo;
 public class DragonFlyInfo extends ItemInfo {
     private int fireballCoolDown;
     private int duration;
-    private int damage;
+    private double damage;
     private double speed;
 
     public int getFireballCoolDown() {
@@ -24,11 +24,11 @@ public class DragonFlyInfo extends ItemInfo {
         this.duration = duration;
     }
 
-    public int getDamage() {
+    public double getDamage() {
         return damage;
     }
 
-    public void setDamage(int damage) {
+    public void setDamage(double damage) {
         this.damage = damage;
     }
