+++ src/main/java/by/dero/gvh/model/itemsinfo/ImprovedBowInfo.java
@@ -3,13 +3,13 @@ package by.dero.gvh.model.itemsinfo;
 import by.dero.gvh.model.ItemInfo;
 
 public class ImprovedBowInfo extends ItemInfo {
-    private int damage;
+    private double damage;
 
-    public int getDamage() {
+    public double getDamage() {
         return damage;
     }
 
-    public void setDamage(int damage) {
+    public void setDamage(double damage) {
         this.damage = damage;
     }
 }
